refactor(hgraph): isolate code slot redirection
